\name{compoisson-package}
\alias{compoisson-package}
\alias{compoisson}
\docType{package}
\title{
	Conway-Maxwell Poisson Distribution
}
\description{
	Provides routines for computing the density of the Conway-Maxwell Poisson
	distribution and fitting parameters to data.
}
\details{
\tabular{ll}{
Package: \tab compoisson\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2008-04-21\cr
License: \tab BSD\cr
}

}
\author{
Jeffrey Dunn

Maintainer: Jeffrey Dunn <jsd115@gmail.com>
}
\references{ Shmueli, G., Minka, T. P., Kadane, J. B., Borle, S. and Boatwright, P., \dQuote{A useful distribution for fitting discrete data: Revival of the Conway-Maxwell-Poisson distribution,} J. Royal Statist. Soc., v54, pp. 127-142, 2005. }
\keyword{ package }
\keyword{ models }
\keyword{ regression }
\seealso{
	See \code{\link{dcom}} for calculating the pmf of the distribution, see
	\code{\link{com.fit}} for fitting parameters.
}
